extracted NavigationBar
