+++ src/components/MenuBar.jsx
@@ -1,44 +1,12 @@
-import ChevronLeftIcon from "@suid/icons-material/ChevronLeft";
-import ChevronRightIcon from "@suid/icons-material/ChevronRight";
-import DynamicFeedIcon from "@suid/icons-material/DynamicFeed";
-import FilterListIcon from "@suid/icons-material/FilterList";
 import MenuIcon from "@suid/icons-material/Menu";
-import ScienceIcon from "@suid/icons-material/Science";
 import AppBar from "@suid/material/AppBar";
 import Container from "@suid/material/Container";
-import Divider from "@suid/material/Divider";
-import Drawer from "@suid/material/Drawer";
 import IconButton from "@suid/material/IconButton";
-import List from "@suid/material/List";
-import ListItem from "@suid/material/ListItem";
-import ListItemIcon from "@suid/material/ListItemIcon";
-import ListItemText from "@suid/material/ListItemText";
-import ListSubheader from "@suid/material/ListSubheader";
 import Slide from "@suid/material/Slide";
-import { useTheme } from "@suid/material/styles";
 import Toolbar from "@suid/material/Toolbar";
-import styled from "@suid/system/styled";
-import { useNavigate } from "solid-app-router";
 import { createEffect, createSignal, mergeProps } from "solid-js";
 import { useScrollTrigger } from "../hooks/useScrollTrigger";
-
-const DrawerHeader = styled("div")(({ theme }) => ({
-  display: "flex",
-  alignItems: "center",
-  padding: theme.spacing(0, 1),
-  // necessary for content to be below app bar
-  // ...theme.mixins.toolbar,
-  justifyContent: "flex-end",
-}));
-
-const StyledDrawer = styled(Drawer)(({ theme }) => ({
-  width: 240,
-  flexShrink: 0,
-  "& .MuiDrawer-paper": {
-    width: 240,
-    boxSizing: "border-box",
-  },
-}));
+import NavigationDrawer from "./NavigationDrawer.jsx";
 
 function HideOnScroll({ enabled, children }) {
   const [direction] = useScrollTrigger();
@@ -61,10 +29,8 @@ function HideOnScroll({ enabled, children }) {
   );
 }
 
-function MenuBar({ hideOnScroll }) {
+function MenuBar({ hideOnScroll = true }) {
   const [open, setOpen] = createSignal(false);
-  const navigate = useNavigate();
-  const theme = useTheme();
 
   function handleDrawerOpen() {
     setOpen(true);
@@ -72,21 +38,6 @@ function MenuBar({ hideOnScroll }) {
 
   function handleDrawerClose() {
     setOpen(false);
-  }
-
-  function navigateHome() {
-    setOpen(false);
-    navigate("/");
-  }
-
-  function navigateSubreddits() {
-    setOpen(false);
-    navigate("/subreddits");
-  }
-
-  function navigateDataset() {
-    setOpen(false);
-    navigate("/dataset");
   }
 
   return (
@@ -106,41 +57,7 @@ function MenuBar({ hideOnScroll }) {
           </Toolbar>
         </AppBar>
       </HideOnScroll>
-      <StyledDrawer variant="persistent" anchor="left" open={open()}>
-        <DrawerHeader>
-          <IconButton onClick={handleDrawerClose}>
-            {theme.direction === "ltr" ? (
-              <ChevronLeftIcon color="primary" />
-            ) : (
-              <ChevronRightIcon color="primary" />
-            )}
-          </IconButton>
-        </DrawerHeader>
-        <Divider />
-        <List subheader={<ListSubheader component="div">News</ListSubheader>}>
-          <ListItem button onClick={navigateHome}>
-            <ListItemIcon>
-              <DynamicFeedIcon color="primary" />
-            </ListItemIcon>
-            <ListItemText primary="Feed" />
-          </ListItem>
-          <ListItem button onClick={navigateSubreddits}>
-            <ListItemIcon>
-              <FilterListIcon color="primary" />
-            </ListItemIcon>
-            <ListItemText primary="Subreddits" />
-          </ListItem>
-        </List>
-        <Divider />
-        <List subheader={<ListSubheader component="div">AI</ListSubheader>}>
-          <ListItem button onClick={navigateDataset}>
-            <ListItemIcon>
-              <ScienceIcon color="primary" />
-            </ListItemIcon>
-            <ListItemText primary="Dataset" />
-          </ListItem>
-        </List>
-      </StyledDrawer>
+      <NavigationDrawer open={open} onClose={handleDrawerClose} />
     </Container>
   );
 }
